+++ docs/source/changelog.rst
@@ -48,6 +48,12 @@ Deprecation notices
 
 * This is the last release to support Python 2.7.
 * This is the last release to support cryptography 2.5 and 2.6.
+* :py:func:`Certificate.objects.init() <django_ca.managers.CertificateManager.init>` has been deprecated in
+  favor of :py:func:`Certificate.objects.create_cert() <django_ca.managers.CertificateManager.create_cert>`.
+  The old method will be removed in ``django-ca==1.16``.
+* :py:func:`~django_ca.profiles.get_cert_profile_kwargs` was only used by
+  :py:func:`~django_ca.managers.CertificateManager.init` and will  thus also be removed in
+  ``django-ca==1.16``.
 * The old format for ``CA_PROFILES`` will be supported until ``django-ca==1.16``.
   Please see :ref:`profiles-pre-114-migration` for migration instructions.